More reliable "wait-until-available" timeout
When edgedb is run in docker and is accessed by a published port,
connection is often accepted by docker itself and then handed over to
the application. In case application is not responding at the moment,
connection will be reset by docker. Since latter happens after
connection is accepted (and therefore after we already have sent a
handshake packet) we have to retry not just `connect()` call but the
whole handshake/auth process.

While this was discovered by working with docker this also covers the
whole range of other potential networking issues when exstablishing a
connection.