Remove IsolationLevel
At the moment, isolation level of a transaction can only be
SERIALIZABLE. It can be specified when opening a connection, but
if it is not, it will default to SERIALIZABLE.

There is also no plans to support other level soon.

IsolationLevel thus serves no purpse and also cannot be set outside
of this crate, since it is not public.

So there is no harm in removing it. When/if we implement other
levels, we can revert this commit.