Support both GEL_ and EDGEDB_ vars (#359)
This allows edgedb-cli to pass all of the shared testcases. To simplify this logic we extract most of the env-var parsing to env.rs. This isn't a full cleanup of this code, but it makes it a little easier to parse what's going on.

Long term we'll need to untangle this error parsing code as it's quite complex and hard to follow.

There's a bit of code here where we need to "lie" that the configuration is complete when we find errors in the environment variables themselves. Ideally this would be done when we surface the errors instead, but that's a more complicated lift.