Fix macOS system validator w/self-signed certs (#387)
When using the system validator on macOS, self-signed certs will fail with -67901 ("The validity period in the certificate exceeds the maximum allowed"). If we have additional roots, we instead want to validate them with the webpki validator. If we have no additional roots, convert that error into UnknownIssuer as no macOS system root should ever issue invalid certs.

We may want to move to the system validator at some point but we'll need to make sure we don't subtly break user installations.